gimp2: tweaks
1. add patch for pthread_threadid_np
2. use legacysupport 1.1 to wrap bins

if we don't use legacysupport 1.1, we run into a problem
with libgcc 7.5+ caused faults, as happens

if we do use legacysupport 1.1, then we have an issue
with the realpath wrapping implementation on in legacysupport
as there is a problem with c++ code that calls <stdlib.h> that
we have not sorted out yet.

The only fix that works for that is to use gcc7 to build gimp2.

Or we might disable the realpath wrapping only using our blocker
for that.